+++ src/Support/Tests/Dummies/DummyAddress.php
@@ -57,4 +57,9 @@ class DummyAddress implements Address
     {
         return $this->address ?? '';
     }
+
+    public function getAddress2(): ?string
+    {
+        return null;
+    }
 }